+++ src/onegov.town/onegov/town/forms/resource.py
@@ -1,5 +1,6 @@
 from onegov.core.utils import sanitize_html
 from onegov.form import Form, with_options
+from onegov.form.validators import ValidFormDefinition
 from onegov.town import _
 from onegov.town.utils import mark_images
 from wtforms import StringField, TextAreaField, validators
@@ -20,6 +21,12 @@ class ResourceForm(Form):
         widget=with_options(TextArea, class_='editor'),
         filters=[sanitize_html, mark_images])
 
+    definition = TextAreaField(
+        label=_(u"Form Definition"),
+        validators=[validators.Optional(), ValidFormDefinition()],
+        widget=with_options(TextArea, rows=32, **{'data-editor': 'form'}),
+    )
+
     def update_model(self, model):
         """ Stores the form values on the page. """
         model.title = self.title.data
@@ -29,9 +36,11 @@ class ResourceForm(Form):
         model.content = {
             'text': self.text.data
         }
+        model.definition = self.definition.data
 
     def apply_model(self, model):
         """ Stores the page values on the form. """
         self.title.data = model.title
         self.lead.data = model.meta.get('lead', '')
         self.text.data = model.content.get('text', '')
+        self.definition.data = model.definition or ""